requirements.txt
bittensor
torch
openai>=1.20.0
loguru
apify-client
elasticsearch
python-dotenv
pydantic
wandb
transformers
tqdm
requests
python-dateutil
